feat: new back and close button components used for header
Keeping the legacy version of the close button, with a new name.
Updated component previously using `HeaderRightClose`
